pyproject.toml
[tool.poetry]
name = "resto"
version = "0.1.0"
description = ""
authors = ["Dominique Eav <dominique.eav@franceimpro.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 200
